91_вынос метода в extension Fragment
